requirements.txt
discord.py>=2.3.2
cloudinary
Flask[async]
aiohttp